attestationreport: refactoring
Moved TPM and SNP specific functions to dedicated
files.
