+++ attestationreport/attestationreport.go
@@ -23,11 +23,9 @@ import (
 	"crypto/rand"
 	"crypto/rsa"
 	"crypto/sha256"
-	"crypto/sha512"
 	"crypto/x509"
 	"encoding/asn1"
 	"encoding/base64"
-	"encoding/binary"
 	"encoding/hex"
 	"encoding/json"
 	"encoding/pem"
@@ -41,8 +39,6 @@ import (
 	"gopkg.in/square/go-jose.v2"
 
 	"time"
-
-	"github.com/google/go-tpm/tpm2"
 )
 
 // Measurement is a generic interface for a Measurement, such as a TpmMeasurement
@@ -330,60 +326,6 @@ type SignerRoles struct {
 	ArSigners          []string
 	ConnDescSigners    []string
 }
-
-type snpreport struct {
-	// Table 24 @ https://www.amd.com/system/files/TechDocs/56860.pdf
-	Status     uint32
-	ReportSize uint32
-	Reserved0  [24]byte
-	// Table 21 @ https://www.amd.com/system/files/TechDocs/56860.pdf
-	Version         uint32
-	GuestSvn        uint32
-	Policy          uint64
-	FamilyId        [16]byte
-	ImageId         [16]byte
-	Vmpl            uint32
-	SignatureAlgo   uint32
-	Currentcb       uint64 // platform_version
-	PlatformInfo    uint64
-	AuthorKeyEn     uint32
-	Reserved1       uint32
-	ReportData      [64]byte
-	Measurement     [48]byte
-	HostData        [32]byte
-	IdKeyDigest     [48]byte
-	AuthorKeyDigest [48]byte
-	ReportId        [32]byte
-	ReportIdMa      [32]byte
-	ReportedTcb     uint64
-	Reserved2       [24]byte
-	ChipId          [64]byte
-	//Reserved3 [192]byte
-	CommittedTcb   uint64
-	CurrentBuild   uint8
-	CurrentMinor   uint8
-	CurrentMajor   uint8
-	Reserved3a     uint8
-	CommittedBuild uint8
-	CommittedMinor uint8
-	CommittedMajor uint8
-	Reserved3b     uint8
-	LaunchTcb      uint64
-	Reserved3c     [168]byte
-	// Table 23 @ https://www.amd.com/system/files/TechDocs/56860.pdf
-	SignatureR [72]byte
-	SignatureS [72]byte
-	Reserved4  [368]byte
-}
-
-const (
-	ecdsa384_with_sha384 = 1
-)
-
-const (
-	header_offset    = 0x20
-	signature_offset = 0x2A0
-)
 
 // Generate generates an attestation report with the provided
 // nonce 'nonce' and manifests and descriptions 'metadata'. The manifests and
@@ -1123,132 +1065,6 @@ func extendVerifications(pcr int, rtmVer, osVer []TpmVerification) ([]byte, int)
 	return result, n
 }
 
-func verifyTpmMeasurements(tpmM *TpmMeasurement, nonce []byte, rtmVerifications, osVerifications []TpmVerification) (TpmMeasurementResult, bool) {
-	result := TpmMeasurementResult{}
-	ok := true
-
-	// Extend the verifications to re-calculate the PCR value and evaluate it against the measured
-	// PCR value. In case of a measurement list, also extend the measured values to re-calculate
-	// the measured PCR value
-	verifications := make(map[int][]byte)
-
-	numExtends := 0
-	for _, hce := range tpmM.HashChain {
-		pcrRes := PcrResult{}
-		pcrRes.Pcr = int(hce.Pcr)
-
-		if len(hce.Sha256) == 1 {
-			// This means, that the value from the PCR was directly taken for the measurement
-			measurement, _ := hex.DecodeString(hce.Sha256[0])
-
-			// Calculate the expected PCR value from the verifications
-			ver, n := extendVerifications(int(hce.Pcr), rtmVerifications, osVerifications)
-			verifications[int(hce.Pcr)] = ver
-			numExtends += n
-
-			// Compare the expected PCR value with the measured PCR value
-			if bytes.Compare(measurement, verifications[int(hce.Pcr)]) == 0 {
-				pcrRes.Validation.Success = true
-			} else {
-				msg := fmt.Sprintf("PCR value did not match expectation: %v vs. %v", hce.Sha256[0], hex.EncodeToString(verifications[int(hce.Pcr)]))
-				pcrRes.Validation.setFalseMulti(&msg)
-				ok = false
-			}
-		} else {
-			// This means, that the values of the individual software artefacts were stored
-			// In this case, we must extend those values in order to re-calculate the final
-			// PCR value
-			hash := make([]byte, 32)
-			for _, sha256 := range hce.Sha256 {
-				h, _ := hex.DecodeString(sha256)
-				hash = extendHash(hash, h)
-
-				// Check, if a verification exists for the measured value
-				v := getVerification(h, rtmVerifications, osVerifications)
-				if v != nil {
-					pcrRes.Validation.Success = true
-					verifications[int(hce.Pcr)] = hash
-					numExtends++
-				} else {
-					msg := fmt.Sprintf("No verification found for measurement: %v", sha256)
-					pcrRes.Validation.setFalseMulti(&msg)
-					ok = false
-				}
-			}
-		}
-		result.PcrRecalculation = append(result.PcrRecalculation, pcrRes)
-	}
-
-	// Check that every verification was extended
-	if numExtends != (len(rtmVerifications) + len(osVerifications)) {
-		msg := fmt.Sprintf("Could not find each expected artefact in measurements (Total: %v, Expected: %v)", numExtends, len(rtmVerifications)+len(osVerifications))
-		result.Summary.setFalse(&msg)
-		ok = false
-	}
-
-	// Extract TPM Quote (TPMS ATTEST) and signature
-	quote, err := hex.DecodeString(tpmM.Message)
-	if err != nil {
-		msg := fmt.Sprintf("Failed to decode TPM quote: %v", err)
-		result.Summary.setFalse(&msg)
-		return result, false
-	}
-	tpmsAttest, err := tpm2.DecodeAttestationData(quote)
-	if err != nil {
-		msg := fmt.Sprintf("Failed to decode TPM attestation data: %v", err)
-		result.Summary.setFalse(&msg)
-		return result, false
-	}
-	sig, err := hex.DecodeString(tpmM.Signature)
-	if err != nil {
-		msg := fmt.Sprintf("Failed to decode TPM signature: %v", err)
-		result.Summary.setFalse(&msg)
-		return result, false
-	}
-
-	// Verify nonce with nonce from TPM Quote
-	if bytes.Compare(nonce, tpmsAttest.ExtraData) == 0 {
-		result.QuoteFreshness.Success = true
-	} else {
-		msg := fmt.Sprintf("Nonces mismatch: Supplied Nonce = %v, TPM Quote Nonce = %v)", hex.EncodeToString(nonce), hex.EncodeToString(tpmsAttest.ExtraData))
-		result.QuoteFreshness.setFalse(&msg)
-		ok = false
-	}
-
-	// Verify aggregated PCR against TPM Quote PCRDigest: Hash all verifications together then compare
-	sum := make([]byte, 0)
-	for i := range tpmM.HashChain {
-		_, ok := verifications[int(tpmM.HashChain[i].Pcr)]
-		if !ok {
-			continue
-		}
-		sum = append(sum, verifications[int(tpmM.HashChain[i].Pcr)]...)
-	}
-	verPcr := sha256.Sum256(sum)
-	if bytes.Compare(verPcr[:], tpmsAttest.AttestedQuoteInfo.PCRDigest) == 0 {
-		result.AggPcrQuoteMatch.Success = true
-	} else {
-		msg := fmt.Sprintf("Aggregated PCR does not match Quote PCR: %v vs. %v", hex.EncodeToString(verPcr[:]), hex.EncodeToString(tpmsAttest.AttestedQuoteInfo.PCRDigest))
-		result.AggPcrQuoteMatch.setFalse(&msg)
-		ok = false
-	}
-
-	result.QuoteSignature = verifyTpmQuoteSignature(quote, sig, tpmM.Certs.Leaf)
-	if result.QuoteSignature.Signature.Success == false {
-		ok = false
-	}
-
-	certCheck, certChainOk := verifyCertChain(&tpmM.Certs)
-	if !certChainOk {
-		ok = false
-	}
-	result.QuoteSignature.CertCheck = certCheck
-
-	result.Summary.Success = ok
-
-	return result, ok
-}
-
 func verifySwMeasurements(swMeasurements []SwMeasurement, appManifests []AppManifest) ([]SwMeasurementResult, bool) {
 	swMeasurementResults := make([]SwMeasurementResult, 0)
 	ok := true
@@ -1358,164 +1174,6 @@ func verifyCertChain(certs *CertChain) (Result, bool) {
 	return result, true
 }
 
-func verifyTpmQuoteSignature(quote, sig []byte, cert string) SignatureResult {
-	result := SignatureResult{}
-	result.Signature.Success = true
-
-	buf := new(bytes.Buffer)
-	buf.Write((sig))
-	tpmtSig, err := tpm2.DecodeSignature(buf)
-	if err != nil {
-		msg := fmt.Sprintf("Failed to decode TPM signature: %v", err)
-		result.Signature.setFalse(&msg)
-		return result
-	}
-
-	if tpmtSig.Alg != tpm2.AlgRSASSA {
-		msg := fmt.Sprintf("Hash algorithm %v not supported", tpmtSig.Alg)
-		result.Signature.setFalse(&msg)
-		return result
-	}
-
-	// Extract public key from x509 certificate
-	x509Cert, err := loadCert([]byte(cert))
-	if err != nil {
-		msg := fmt.Sprintf("Failed to parse TPM certificate: %v", err)
-		result.Signature.setFalse(&msg)
-		return result
-	}
-	pubKey, ok := x509Cert.PublicKey.(*rsa.PublicKey)
-	if !ok {
-		msg := fmt.Sprintf("Failed to extract public key from certificate")
-		result.Signature.setFalse(&msg)
-		return result
-	}
-	hashAlg, err := tpmtSig.RSA.HashAlg.Hash()
-	if err != nil {
-		msg := fmt.Sprintf("Hash algorithm not supported")
-		result.Signature.setFalse(&msg)
-		return result
-	}
-
-	// Store Name and Organization of AK Cert for the report
-	result.Name = x509Cert.Subject.CommonName
-	result.Organization = x509Cert.Subject.Organization
-
-	// Hash the quote and Verify the TPM Quote signature
-	hashed := sha256.Sum256(quote)
-	err = rsa.VerifyPKCS1v15(pubKey, hashAlg, hashed[:], tpmtSig.RSA.Signature)
-	if err != nil {
-		msg := fmt.Sprintf("Failed to verify TPM quote signature: %v", err)
-		result.Signature.setFalse(&msg)
-		return result
-	}
-	return result
-}
-
-func verifySnpMeasurements(snpM *SnpMeasurement, v *SnpVerification, nonce []byte) (SnpMeasurementResult, bool) {
-	result := SnpMeasurementResult{}
-	ok := true
-
-	var s snpreport
-	b := bytes.NewBuffer(snpM.Report)
-
-	err := binary.Read(b, binary.LittleEndian, &s)
-	if err != nil {
-		msg := fmt.Sprintf("Failed to decode SNP attestation data: %v", err)
-		result.Summary.setFalse(&msg)
-		return result, false
-	}
-
-	// Compare Nonce for Freshness (called Report Data in the SNP Attestation Report Structure)
-	if cmp := bytes.Compare(s.ReportData[:], nonce); cmp != 0 {
-		msg := fmt.Sprintf("Nonces mismatch: Supplied Nonce = %v, TPM Quote Nonce = %v)", hex.EncodeToString(nonce), hex.EncodeToString(s.ReportData[:]))
-		result.Freshness.setFalse(&msg)
-		ok = false
-	}
-
-	// Verify Signature, created with SNP VCEK private key
-	sig, ret := verifySnpSignature(snpM.Report, s, []byte(snpM.Certs.Leaf))
-	if !ret {
-		ok = false
-	}
-	result.Signature = sig
-
-	// Verify the SNP certificate chain
-	certCheck, certChainOk := verifyCertChain(&snpM.Certs)
-	if !certChainOk {
-		ok = false
-	}
-	result.Signature.CertCheck = certCheck
-
-	// Compare Measurements
-	m, err := hex.DecodeString(v.Sha256)
-	if err != nil {
-		msg := fmt.Sprintf("Failed to decode SNP Measurement: %v", err)
-		result.Summary.setFalse(&msg)
-		ok = false
-	}
-	if cmp := bytes.Compare(s.Measurement[:], m); cmp != 0 {
-		msg := fmt.Sprintf("SNP Measurement mismatch: Supplied measurement = %v, SNP report measurement = %v", v.Sha256, hex.EncodeToString(s.Measurement[:]))
-		result.MeasurementMatch.setFalse(&msg)
-		ok = false
-	}
-
-	// Compare SNP parameters
-	result.ParamsMatch, ret = verifySnpParams(&s, v)
-	if !ret {
-		ok = false
-	}
-
-	result.Summary.Success = ok
-
-	return result, ok
-}
-
-func verifySnpParams(s *snpreport, v *SnpVerification) (Result, bool) {
-	result := Result{
-		Success: true,
-	}
-
-	if s.Version != v.Version {
-		msg := fmt.Sprintf("SNP report version mismatch: Report = %v, supplied = %v", s.Version, v.Version)
-		result.setFalse(&msg)
-	}
-
-	abiMajor := uint8(s.Policy & 0xFF)
-	abiMinor := uint8((s.Policy >> 8) & 0xFF)
-	smt := (s.Policy & (1 << 16)) != 0
-	migration := (s.Policy & (1 << 18)) != 0
-	debug := (s.Policy & (1 << 19)) != 0
-	singleSocket := (s.Policy & (1 << 20)) != 0
-
-	if abiMajor != v.Policy.AbiMajor {
-		msg := fmt.Sprintf("SNP report AbiMinor mismatch: Report = %v, supplied = %v", abiMajor, v.Policy.AbiMajor)
-		result.setFalse(&msg)
-	}
-	if abiMinor != v.Policy.AbiMinor {
-		msg := fmt.Sprintf("SNP report AbiMinor mismatch: Report = %v, supplied = %v", abiMinor, v.Policy.AbiMinor)
-		result.setFalse(&msg)
-	}
-	if smt != v.Policy.Smt {
-		msg := fmt.Sprintf("SNP report SMT mismatch: Report = %v, supplied = %v", smt, v.Policy.Smt)
-		result.setFalse(&msg)
-	}
-	if migration != v.Policy.Migration {
-		msg := fmt.Sprintf("SNP report Migration Agent mismatch: Report = %v, supplied = %v", migration, v.Policy.Migration)
-		result.setFalse(&msg)
-	}
-	if debug != v.Policy.Debug {
-		msg := fmt.Sprintf("SNP report Debug Support mismatch: Report = %v, supplied = %v", debug, v.Policy.Debug)
-		result.setFalse(&msg)
-	}
-	if singleSocket != v.Policy.SingleSocket {
-		msg := fmt.Sprintf("SNP report SingleSocket mismatch: Report = %v, supplied = %v", singleSocket, v.Policy.SingleSocket)
-		result.setFalse(&msg)
-	}
-
-	return result, result.Success
-}
-
 func checkExtensionUint8(cert *x509.Certificate, oid string, value uint8) error {
 
 	for _, ext := range cert.Extensions {
@@ -1553,119 +1211,6 @@ func checkExtensionBuf(cert *x509.Certificate, oid string, buf []byte) error {
 	}
 
 	return fmt.Errorf("Extension %v not present in certificate", oid)
-}
-
-func checkExtensions(cert *x509.Certificate, report *snpreport) (ResultMulti, bool) {
-	result := ResultMulti{}
-	ok := true
-	tcb := report.ReportedTcb
-
-	if err := checkExtensionUint8(cert, "1.3.6.1.4.1.3704.1.3.1", uint8(tcb)); err != nil {
-		msg := fmt.Sprintf("SEV BL Extension Check failed: %v", err)
-		result.setFalseMulti(&msg)
-		ok = false
-	}
-
-	if err := checkExtensionUint8(cert, "1.3.6.1.4.1.3704.1.3.2", uint8(tcb>>8)); err != nil {
-		msg := fmt.Sprintf("SEV TEE Extension Check failed: %v", err)
-		result.setFalseMulti(&msg)
-		ok = false
-	}
-
-	if err := checkExtensionUint8(cert, "1.3.6.1.4.1.3704.1.3.3", uint8(tcb>>48)); err != nil {
-		msg := fmt.Sprintf("SEV SNP Extension Check failed: %v", err)
-		result.setFalseMulti(&msg)
-		ok = false
-	}
-
-	if err := checkExtensionUint8(cert, "1.3.6.1.4.1.3704.1.3.8", uint8(tcb>>56)); err != nil {
-		msg := fmt.Sprintf("SEV UCODE Extension Check failed: %v", err)
-		result.setFalseMulti(&msg)
-		ok = false
-	}
-
-	if err := checkExtensionBuf(cert, "1.3.6.1.4.1.3704.1.4", report.ChipId[:]); err != nil {
-		msg := fmt.Sprintf("Chip ID Extension Check failed: %v", err)
-		result.setFalseMulti(&msg)
-		ok = false
-	}
-
-	result.Success = ok
-
-	return result, ok
-}
-
-func verifySnpSignature(reportRaw []byte, report snpreport, cert []byte) (SignatureResult, bool) {
-	result := SignatureResult{}
-
-	if len(reportRaw) < (header_offset + signature_offset) {
-		msg := fmt.Sprintf("Internal Error: Report buffer too small")
-		result.Signature.setFalse(&msg)
-		return result, false
-	}
-
-	// Strip the header and the signature from the report and hash for signature verification
-	// Table 21, 23 @ https://www.amd.com/system/files/TechDocs/56860.pdf
-	digest := sha512.Sum384(reportRaw[0x20 : 0x20+0x2A0])
-
-	// Golang SetBytes expects BigEndian byte array, but SNP values are little endian
-	rRaw := report.SignatureR[:]
-	for i := 0; i < len(rRaw)/2; i++ {
-		rRaw[i], rRaw[len(rRaw)-i-1] = rRaw[len(rRaw)-i-1], rRaw[i]
-	}
-	sRaw := report.SignatureS[:]
-	for i := 0; i < len(sRaw)/2; i++ {
-		sRaw[i], sRaw[len(sRaw)-i-1] = sRaw[len(sRaw)-i-1], sRaw[i]
-	}
-
-	// Convert r, s to Big Int
-	r := new(big.Int)
-	r.SetBytes(rRaw)
-	s := new(big.Int)
-	s.SetBytes(sRaw)
-
-	// Load the VCEK certificate
-	c, err := loadCert(cert)
-	if err != nil {
-		msg := fmt.Sprintf("Failed to load certificate: %v", err)
-		result.Signature.setFalse(&msg)
-		return result, false
-	}
-	result.Name = c.Subject.CommonName
-	result.Organization = c.Subject.Organization
-
-	// Examine SNP x509 extensions
-	extensionResult, ok := checkExtensions(c, &report)
-	result.ExtensionsCheck = &extensionResult
-	if !ok {
-		return result, false
-	}
-
-	// Check that the algorithm is supported
-	if report.SignatureAlgo != ecdsa384_with_sha384 {
-		msg := fmt.Sprintf("Siganture Algorithm %v not supported", report.SignatureAlgo)
-		result.Signature.setFalse(&msg)
-		return result, false
-	}
-
-	// Extract the public key from the certificate
-	pub, ok := c.PublicKey.(*ecdsa.PublicKey)
-	if !ok {
-		msg := fmt.Sprintf("Failed to extract ECDSA public key from certificate")
-		result.Signature.setFalse(&msg)
-		return result, false
-	}
-
-	// Verify ECDSA Signature represented by r and s
-	ok = ecdsa.Verify(pub, digest[:], r, s)
-	if !ok {
-		msg := fmt.Sprintf("Failed to verify SNP report signature")
-		result.Signature.setFalse(&msg)
-		return result, false
-	}
-	log.Trace("Successfully verified SNP report signature")
-
-	return result, true
 }
 
 // Used for the JOSE Opaque Signer Interface. This enables signing
